requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.5
Markdown==2.6.11
numpy==1.14.1
opencv-python==3.4.0.12
pandas==0.21.0
Pillow==5.0.0
protobuf==3.5.2
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
tqdm==4.11.2
Werkzeug==0.14.1
